+++ src/subject.test.ts
@@ -22,6 +22,11 @@ const mockSubjectWithIntegerLiteral = 'https://subject8.com/';
 const mockSubjectWithDecimalLiteral = 'https://subject9.com/';
 const mockSubjectWithDifferentTypesOfLiterals = 'https://subject10.com/';
 const mockSubjectWithDifferentPredicates = 'https://subject11.com/';
+const mockSubjectWithLocaleStringLiteral = 'https://subject12.com/';
+const mockSubjectWithLocaleAndRegularStringLiteral = 'https://subject13.com/';
+const mockSubjectWithMultipleLocaleStringLiterals = 'https://subject14.com/';
+const mockSubjectWithLangStringLiteral = 'https://subject15.com/';
+const mockSubjectWithMultipleSameLocaleStringLiterals = 'https://subject16.com/';
 const mockTypedSubject = 'https://subject7.com/';
 const mockEmptySubject = 'https://empty-subject.com/';
 const mockPredicate = 'https://mock-predicate.com/';
@@ -40,6 +45,14 @@ const mockLiteralInteger = 1337;
 const mockObjectIntegerLiteral = literal(mockLiteralInteger);
 const mockLiteralDecimal = 4.2;
 const mockObjectDecimalLiteral = literal(mockLiteralDecimal.toString(), namedNode('http://www.w3.org/2001/XMLSchema#decimal'));
+const mockLiteralLocaleString = 'Dutch (NL)';
+const mockLocale = 'nl-NL';
+const mockObjectLocaleStringLiteral = literal(mockLiteralLocaleString, mockLocale);
+const mockLiteralLocaleString2 = 'Dutch (NL) 2';
+const mockObjectLocaleStringLiteral2 = literal(mockLiteralLocaleString2, mockLocale);
+const mockLiteralLangString = 'Dutch (no country specified)';
+const mockLang = 'nl';
+const mockObjectLangStringLiteral = literal(mockLiteralLangString, mockLang);
 const mockTypeObject = 'https://mock-type-object.com/';
 const mockBlankNode = 'arbitrary-blank-node-id';
 const mockTriples = [
@@ -72,6 +85,13 @@ const mockTriples = [
   triple(namedNode(mockSubjectWithDifferentPredicates), namedNode(mockPredicate), mockObjectLiteral),
   triple(namedNode(mockSubjectWithDifferentPredicates), namedNode(mockPredicate), namedNode(mockObjectRef)),
   triple(namedNode(mockSubjectWithDifferentPredicates), namedNode(mockPredicate2), namedNode(mockObjectRef2)),
+  triple(namedNode(mockSubjectWithLocaleStringLiteral), namedNode(mockPredicate), mockObjectLocaleStringLiteral),
+  triple(namedNode(mockSubjectWithLocaleAndRegularStringLiteral), namedNode(mockPredicate), mockObjectLiteral),
+  triple(namedNode(mockSubjectWithLocaleAndRegularStringLiteral), namedNode(mockPredicate), mockObjectLocaleStringLiteral),
+  triple(namedNode(mockSubjectWithMultipleLocaleStringLiterals), namedNode(mockPredicate), mockObjectLangStringLiteral),
+  triple(namedNode(mockSubjectWithMultipleLocaleStringLiterals), namedNode(mockPredicate), mockObjectLocaleStringLiteral),
+  triple(namedNode(mockSubjectWithMultipleSameLocaleStringLiterals), namedNode(mockPredicate), mockObjectLocaleStringLiteral),
+  triple(namedNode(mockSubjectWithMultipleSameLocaleStringLiterals), namedNode(mockPredicate), mockObjectLocaleStringLiteral2),
 ];
 const turtle = triplesToTurtle(mockTriples);
 jest.mock('./store', () => ({
@@ -197,7 +217,6 @@ describe('getString', () => {
       .toBe(mockLiteralValue);
   });
 
-
   it('should return null if a Reference is found instead of a string Literal', async () => {
     const mockTripleDocument = await getMockTripleDocument();
     const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
@@ -209,6 +228,64 @@ describe('getString', () => {
     const mockTripleDocument = await getMockTripleDocument();
     const subject = initialiseSubject(mockTripleDocument, mockSubjectWithDateLiteral);
     expect(subject.getString(mockPredicate))
+      .toBeNull();
+  });
+
+  it('should return null if a locale string Literal is found', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLocaleStringLiteral);
+    expect(subject.getString(mockPredicate))
+      .toBeNull();
+  });
+});
+
+describe('getLocaleString', () => {
+  it('should return a found locale string Literal', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLocaleStringLiteral);
+    expect(subject.getLocaleString(mockPredicate, mockLocale))
+      .toBe(mockLiteralLocaleString);
+  });
+
+  it('should return the desired locale string Literal when multiple are present', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithMultipleLocaleStringLiterals);
+    expect(subject.getLocaleString(mockPredicate, mockLocale))
+      .toBe(mockLiteralLocaleString);
+  });
+
+  it('should return the desired locale string Literal when regular strings are present as well', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLocaleAndRegularStringLiteral);
+    expect(subject.getLocaleString(mockPredicate, mockLocale))
+      .toBe(mockLiteralLocaleString);
+  });
+
+  it('should return null if the language matches but the full locale does not - picking the right one is left to the developer', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLocaleStringLiteral);
+    expect(subject.getLocaleString(mockPredicate, mockLang))
+      .toBeNull();
+  });
+
+  it('should return null if the language matches but there is no info about the locale - it is up to developer to decide whether to look for those as well', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLangStringLiteral);
+    expect(subject.getLocaleString(mockPredicate, mockLocale))
+      .toBeNull();
+  });
+
+  it('should return null if a non-locale-string Literal is found', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithDateLiteral);
+    expect(subject.getLocaleString(mockPredicate, mockLocale))
+      .toBeNull();
+  });
+
+  it('should return null if a non-locale string Literal is found', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    expect(subject.getLocaleString(mockPredicate, mockLocale))
       .toBeNull();
   });
 });
@@ -297,6 +374,22 @@ describe('getAllStrings', () => {
     const mockTripleDocument = await getMockTripleDocument();
     const subject = initialiseSubject(mockTripleDocument, mockEmptySubject);
     expect(subject.getAllStrings(mockPredicate))
+      .toEqual([]);
+  });
+});
+
+describe('getAllLocaleStrings', () => {
+  it('should only return locale string Literals that match the given locale', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithMultipleSameLocaleStringLiterals);
+    expect(subject.getAllLocaleStrings(mockPredicate, mockLocale))
+      .toEqual([mockLiteralLocaleString, mockLiteralLocaleString2]);
+  });
+
+  it('should return an empty array if nothing is found', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockEmptySubject);
+    expect(subject.getAllLocaleStrings(mockPredicate, mockLocale))
       .toEqual([]);
   });
 });
@@ -543,6 +636,136 @@ describe('addLiteral', () => {
   });
 });
 
+describe('addString', () => {
+  it('should produce Triples that the Document can store in the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    subject.addString(mockPredicate, 'Some string value');
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#string');
+    expect(pendingAdditions[0].object.value).toBe('Some string value');
+  });
+
+  it('should throw an error when something other than a string was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    expect(() => subject.addString(mockPredicate, 1337 as any))
+      .toThrowError('The given value is not a string.');
+  });
+});
+
+describe('addLocaleString', () => {
+  it('should produce Triples that the Document can store in the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    subject.addLocaleString(mockPredicate, 'Some localised string value', 'en-GB');
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/1999/02/22-rdf-syntax-ns#langString');
+    expect((pendingAdditions[0].object as Literal).language)
+      .toBe('en-gb');
+    expect(pendingAdditions[0].object.value).toBe('Some localised string value');
+  });
+
+  it('should throw an error when something other than a string was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    expect(() => subject.addLocaleString(mockPredicate, 1337 as any, 'nl-NL'))
+      .toThrowError('The given value is not a string.');
+  });
+});
+
+describe('addInteger', () => {
+  it('should produce Triples that the Document can store in the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    subject.addInteger(mockPredicate, 42);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#integer');
+    expect(pendingAdditions[0].object.value).toBe('42');
+  });
+
+  it('should throw an error when something other than a number was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    expect(() => subject.addInteger(mockPredicate, new Date() as any))
+      .toThrowError('The given value is not an integer.');
+  });
+
+  it('should throw an error when a decimal number was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    expect(() => subject.addInteger(mockPredicate, 13.37))
+      .toThrowError('The given value is not an integer.');
+  });
+});
+
+describe('addDecimal', () => {
+  it('should produce Triples that the Document can store in the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    subject.addDecimal(mockPredicate, 13.37);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#decimal');
+    expect(pendingAdditions[0].object.value).toBe('13.37');
+  });
+
+  it('should store an integer as a decimal number', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    subject.addDecimal(mockPredicate, 42);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect(pendingAdditions[0].object.value).toBe('42');
+  });
+
+  it('should throw an error when something other than a number was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    expect(() => subject.addDecimal(mockPredicate, 'Not a number' as any))
+      .toThrowError('The given value is not a decimal.');
+  });
+});
+
+describe('addDateTime', () => {
+  it('should produce Triples that the Document can store in the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    subject.addDateTime(mockPredicate, new Date(Date.UTC(1970, 0)));
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#dateTime');
+    expect(pendingAdditions[0].object.value).toBe('1970-01-01T00:00:00Z');
+  });
+
+  it('should throw an error when something other than a Date was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
+    expect(() => subject.addDateTime(mockPredicate, 'Not a Date' as any))
+      .toThrowError('The given value is not a DateTime.');
+  });
+});
+
 describe('removeLiteral', () => {
   it('should produce Triples that the Document can apply to the user\'s Pod', async () => {
     const mockTripleDocument = await getMockTripleDocument();
@@ -595,6 +818,137 @@ describe('removeLiteral', () => {
   });
 });
 
+describe('removeString', () => {
+  it('should produce Triples that the Document can apply to the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithTwoLiterals);
+    subject.removeString(mockPredicate, 'Some string');
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingAdditions).toEqual([]);
+    expect(pendingDeletions.length).toBe(1);
+    expect(pendingDeletions[0].object.termType).toBe('Literal');
+    expect((pendingDeletions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#string');
+    expect(pendingDeletions[0].object.value).toBe('Some string');
+  });
+
+  it('should throw an error if something other than a string was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithIntegerLiteral);
+    expect(() => subject.removeString(mockPredicate, new Date() as any))
+      .toThrowError('The given value is not a string.');
+  });
+});
+
+describe('removeLocaleString', () => {
+  it('should produce Triples that the Document can apply to the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithTwoLiterals);
+    subject.removeLocaleString(mockPredicate, 'Some string', 'en-GB');
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingAdditions).toEqual([]);
+    expect(pendingDeletions.length).toBe(1);
+    expect(pendingDeletions[0].object.termType).toBe('Literal');
+    expect((pendingDeletions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/1999/02/22-rdf-syntax-ns#langString');
+    expect((pendingDeletions[0].object as Literal).language).toBe('en-gb');
+    expect(pendingDeletions[0].object.value).toBe('Some string');
+  });
+
+  it('should throw an error if something other than a string was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithIntegerLiteral);
+    expect(() => subject.removeLocaleString(mockPredicate, 42 as any, 'nl-NL'))
+      .toThrowError('The given value is not a string.');
+  });
+});
+
+describe('removeInteger', () => {
+  it('should produce Triples that the Document can apply to the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithIntegerLiteral);
+    subject.removeInteger(mockPredicate, 42);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingAdditions).toEqual([]);
+    expect(pendingDeletions.length).toBe(1);
+    expect(pendingDeletions[0].object.termType).toBe('Literal');
+    expect((pendingDeletions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#integer');
+    expect(pendingDeletions[0].object.value).toBe('42');
+  });
+
+  it('should throw an error if something other than a number was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithIntegerLiteral);
+    expect(() => subject.removeInteger(mockPredicate, 'Not a number' as any))
+      .toThrowError('The given value is not an integer.');
+  });
+
+  it('should throw an error if a non-integer number was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithIntegerLiteral);
+    expect(() => subject.removeInteger(mockPredicate, 13.37))
+      .toThrowError('The given value is not an integer.');
+  });
+});
+
+describe('removeDecimal', () => {
+  it('should produce Triples that the Document can apply to the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithDecimalLiteral);
+    subject.removeDecimal(mockPredicate, 13.37);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingAdditions).toEqual([]);
+    expect(pendingDeletions.length).toBe(1);
+    expect(pendingDeletions[0].object.termType).toBe('Literal');
+    expect((pendingDeletions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#decimal');
+    expect(pendingDeletions[0].object.value).toBe('13.37');
+  });
+
+  it('should store integers as decimals', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithDecimalLiteral);
+    subject.removeDecimal(mockPredicate, 42);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingAdditions).toEqual([]);
+    expect(pendingDeletions.length).toBe(1);
+    expect(pendingDeletions[0].object.termType).toBe('Literal');
+    expect((pendingDeletions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#decimal');
+    expect(pendingDeletions[0].object.value).toBe('42');
+  });
+
+  it('should throw an error if something other than a number was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithDecimalLiteral);
+    expect(() => subject.removeDecimal(mockPredicate, 'Not a number' as any))
+      .toThrowError('The given value is not a decimal.');
+  });
+});
+
+describe('removeDateTime', () => {
+  it('should produce Triples that the Document can apply to the user\'s Pod', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithDateLiteral);
+    subject.removeDateTime(mockPredicate, new Date(Date.UTC(1970, 0)));
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingAdditions).toEqual([]);
+    expect(pendingDeletions.length).toBe(1);
+    expect(pendingDeletions[0].object.termType).toBe('Literal');
+    expect((pendingDeletions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#dateTime');
+    expect(pendingDeletions[0].object.value).toBe('1970-01-01T00:00:00Z');
+  });
+
+  it('should throw an error if something other than a DateTime was given', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithDateLiteral);
+    expect(() => subject.removeDateTime(mockPredicate, 42 as any))
+      .toThrowError('The given value is not a DateTime.');
+  });
+});
+
 describe('setLiteral', () => {
   it('should remove all existing values, whether Literal or Reference', async () => {
     const mockTripleDocument = await getMockTripleDocument();
@@ -616,6 +970,137 @@ describe('setLiteral', () => {
     expect(pendingAdditions.length).toBe(1);
     expect(pendingAdditions[0].object.termType).toBe('Literal');
     expect(pendingAdditions[0].object.value).toBe(mockLiteralValue2);
+  });
+});
+
+describe('setString', () => {
+  it('should remove all existing values, whether Literal or Reference', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
+    subject.setString(mockPredicate, 'Some string');
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        mockObjectLiteral,
+      ),
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        namedNode(mockObjectRef),
+      ),
+    ]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#string');
+    expect(pendingAdditions[0].object.value).toBe('Some string');
+  });
+});
+
+describe('setLocaleString', () => {
+  it('should remove all existing values, whether Literal or Reference', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
+    subject.setLocaleString(mockPredicate, 'Some string', 'en-GB');
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        mockObjectLiteral,
+      ),
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        namedNode(mockObjectRef),
+      ),
+    ]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/1999/02/22-rdf-syntax-ns#langString');
+    expect((pendingAdditions[0].object as Literal).language).toBe('en-gb');
+    expect(pendingAdditions[0].object.value).toBe('Some string');
+  });
+});
+
+describe('setInteger', () => {
+  it('should remove all existing values, whether Literal or Reference', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
+    subject.setInteger(mockPredicate, 42);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        mockObjectLiteral,
+      ),
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        namedNode(mockObjectRef),
+      ),
+    ]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#integer');
+    expect(pendingAdditions[0].object.value).toBe('42');
+  });
+});
+
+describe('setDecimal', () => {
+  it('should remove all existing values, whether Literal or Reference', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
+    subject.setDecimal(mockPredicate, 13.37);
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        mockObjectLiteral,
+      ),
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        namedNode(mockObjectRef),
+      ),
+    ]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#decimal');
+    expect(pendingAdditions[0].object.value).toBe('13.37');
+  });
+});
+
+describe('setDateTime', () => {
+  it('should remove all existing values, whether Literal or Reference', async () => {
+    const mockTripleDocument = await getMockTripleDocument();
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
+    subject.setDateTime(mockPredicate, new Date(Date.UTC(1970, 0)));
+    const [pendingDeletions, pendingAdditions] = subject.getPendingTriples();
+    expect(pendingDeletions).toEqual([
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        mockObjectLiteral,
+      ),
+      triple(
+        namedNode(mockSubjectWithLiteralThenRef),
+        namedNode(mockPredicate),
+        namedNode(mockObjectRef),
+      ),
+    ]);
+    expect(pendingAdditions.length).toBe(1);
+    expect(pendingAdditions[0].object.termType).toBe('Literal');
+    expect((pendingAdditions[0].object as Literal).datatype.value)
+      .toBe('http://www.w3.org/2001/XMLSchema#dateTime');
+    expect(pendingAdditions[0].object.value).toBe('1970-01-01T00:00:00Z');
   });
 });